pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "causalbootstrapping"
version = "0.2.1"
description = "CausalBootstrapping is an easy-access implementation and extention of causal bootstrapping (CB) technique for causal analysis. With certain input of observational data, causal graph and variable distributions, CB resamples the data by adjusting the variable distributions which follow intended causal effects."
requires-python = ">=3.9"
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [{name="Jianqiao Mao", email="jxm1417@student.bham.ac.uk"}]
urls = {"Homepage" = "https://github.com/JianqiaoMao/CausalBootstrapping"}
dependencies = [
  'numpy ~= 1.25.0', 
  'grapl-causal == 1.6.1', 
  'scipy ~= 1.13.1', 
  'graphviz == 0.20.3',
  'sympy ~= 1.13.1'
]

[tool.setuptools.packages.find]
include = ["causalbootstrapping*"]
